pyproject.toml
[tool.poetry]
name = "prospectgeo"
version = "0.1.0"
description = "Location-Based Prospect Qualification System"
authors = ["nadavedri <nadav052433@gmail.com>"]
packages = [{ include = "prospectgeo", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.9"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.0"
tenacity = "^9.1.2"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
ruff = "^0.11.8"
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
lint.select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
lint.ignore = []
line-length = 88
target-version = "py312"
src = ["src", "tests"]
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "__pypackages__",
    "dist",
    "build"
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.isort]
known-third-party = ["pytest"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.poetry.scripts]
prospectgeo = "your_project.main:main"